fix: 修复CI编译错误 — MediaPipe API路径、TFLite类型、未使用导入
- MediaPipeFaceDetector: 修复BaseOptions导入路径(vision.core→core)、FaceLandmarkerOptions→FaceLandmarker.FaceLandmarkerOptions、faceBlendshapes()返回Optional类型处理、mapNotNull→filter+map
- MediaPipeFaceAnalyzer: mapNotNull→filter+map修复Java List上的Kotlin扩展问题
- FacePreviewAnalyzer: computeBlendshapeScore参数类型从Classifications→Category
- TongueEnhancer: 添加abs导入、修复times操作符重载歧义
- Pulse模块: 移除未使用的Complex导入(HrvAnalyzer/SanJiaoSimulator/RppGProcessor)

+++ module_tongue/src/main/java/com/bianque/health/tongue/data/TongueEnhancer.kt
@@ -11,6 +11,7 @@ import java.nio.ByteBuffer
 import java.nio.ByteOrder
 import javax.inject.Inject
 import javax.inject.Singleton
+import kotlin.math.abs
 import kotlin.math.max
 import kotlin.math.min
 
@@ -312,9 +313,9 @@ class TongueEnhancer @Inject constructor() {
                 val v01 = tileLuts[d][a][pixelValue]
                 val v11 = tileLuts[d][b][pixelValue]
 
-                val top = v00 + ((v10 - v00) * weightX + 0.5f).toInt()
-                val bottom = v01 + ((v11 - v01) * weightX + 0.5f).toInt()
-                val interpolated = top + ((bottom - top) * weightY + 0.5f).toInt()
+                val top = v00 + ((v10 - v00).toFloat() * weightX + 0.5f).toInt()
+                val bottom = v01 + ((v11 - v01).toFloat() * weightX + 0.5f).toInt()
+                val interpolated = top + ((bottom - top).toFloat() * weightY + 0.5f).toInt()
 
                 result[idx] = interpolated.coerceIn(0, 255)
             }
